arslan thursday work

+++ app/Http/Controllers/PortfolioController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\PortfolioImages;
 use App\Models\Portfolio;
 use Illuminate\Http\Request;
 
@@ -23,17 +24,37 @@ class PortfolioController extends Controller
             'name' => 'required',
             'info' => 'required',
             'type' => 'required',
+            'category' => 'required',
+            'client' => 'required',
+            'project_date' => 'required',
+            'project_url' => 'required',
+            'description' => 'required',
             'image' => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
         ]);
         $input = $request->all();
+        unset($input['image2[]']);
+        $check ="";
         if ($image = $request->file('image')) {
             $destinationPath = 'public/image/portfolio/';
             $profileImage = date('YmdHis') . "." . $image->getClientOriginalExtension();
             $image->move($destinationPath, $profileImage);
             $input['image'] = "$profileImage";
+            $check = "$profileImage";
         }
-    
         Portfolio::create($input);
+        $portfolio = Portfolio::where('image','=', $input['image'])->first();
+        $id = $portfolio->id;
+        if ($image2 = $request->file('image2')) {
+            foreach($image2 as $images) {
+                $portfolioImg = new PortfolioImages();
+                $destinationPath = 'public/image/portfolio/';
+                $profileImage = $images->getClientOriginalName();
+                $images->move($destinationPath, $profileImage);
+                $portfolioImg->portfolio_img = "$profileImage";
+                $portfolioImg->portfolio_id = $id;
+                $portfolioImg->save();
+            }
+        }
         return redirect('portfolio')->with('success','Portfolio created successfully.');
     }
     public function updatePortfolio(Request $request)
@@ -42,14 +63,25 @@ class PortfolioController extends Controller
             'name' => 'required',
             'info' => 'required',
             'type' => 'required',
+            'category' => 'required',
+            'client' => 'required',
+            'project_date' => 'required',
+            'project_url' => 'required',
+            'description' => 'required',
         ]);
         $portfolio = Portfolio::find($request->id);
+        
         $portfolio->name = $request->name;
         $portfolio->info = $request->info;
         $portfolio->type = $request->type;
+        $portfolio->category = $request->category;
+        $portfolio->client = $request->client;
+        $portfolio->project_date = $request->project_date;
+        $portfolio->project_url = $request->project_url;
+        $portfolio->description = $request->description;
         if ($image = $request->file('image')) {
             $destinationPath = 'public/image/portfolio/';
-            $profileImage = date('YmdHis') . "." . $image->getClientOriginalExtension();
+            $profileImage = $image->getClientOriginalName();
             $image->move($destinationPath, $profileImage);
             $portfolio->image = "$profileImage";
         }
@@ -58,25 +90,28 @@ class PortfolioController extends Controller
     }
     public function getPortfolioById($id)
     {
-        $data = [];
+        // $data = [];
         $portfolio = Portfolio::find($id);
-        // $data['id'] = $portfolio->id;
-        // $data['name'] = $portfolio->name;
-        // $data['info'] = $portfolio->info;
-        // $data['type'] = $portfolio->type;
-        // $data['image'] = $portfolio->image;
-        // // echo $data['name'];
-        // echo $data['info'];
-        // echo $data['image'];
-        // dd($portfolio);
-        // die();
-        return view('dashboard/editPortfolio',compact('portfolio'));
+        $portfolioimg = PortfolioImages::where('portfolio_id','=', $portfolio->id)->get();
+
+        return view('dashboard/editPortfolio',compact('portfolio','portfolioimg'));
     }
 
     public function deletePortfolio($id)
     {
         $portfolio = Portfolio::find($id);
-        $portfolio->delete();
+        $image_path = public_path("\image\portfolio\\").$portfolio->image;
+        if (file_exists($image_path)) {
+            @unlink($image_path);
+            $portfolioimg = PortfolioImages::where('portfolio_id','=', $portfolio->id)->get();
+            if ($portfolioimg) {
+                foreach($portfolioimg as $images) {
+                    @unlink(public_path("\image\portfolio\\").$images->portfolio_img);
+                    $images->delete();
+                }
+            }
+            $portfolio->delete();
+        }
         return response()->json(['success'=>'Record has been deleted']);
     }
 }
